+++ packages/react-icons/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [4.7.20](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-icons@4.7.19...@patternfly/react-icons@4.7.20) (2020-12-07)
+
+**Note:** Version bump only for package @patternfly/react-icons
+
+
+
+
+
 ## [4.7.19](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-icons@4.7.18...@patternfly/react-icons@4.7.19) (2020-11-30)
 
 **Note:** Version bump only for package @patternfly/react-icons