+++ packages/react-styles/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [4.7.19](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-styles@4.7.18...@patternfly/react-styles@4.7.19) (2020-12-07)
+
+**Note:** Version bump only for package @patternfly/react-styles
+
+
+
+
+
 ## [4.7.18](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-styles@4.7.17...@patternfly/react-styles@4.7.18) (2020-12-03)